# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil retornado com sucesso."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter perfil do usuário autenticado",
        "tags": [
          "Users - App"
        ]
      }
    },
    "/users/admin": {
      "get": {
        "operationId": "UsersController_getAdminData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Acesso de administrador liberado."
          },
          "403": {
            "description": "Acesso negado (requer role ADMIN)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rota exclusiva para administradores",
        "tags": [
          "Users - App"
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "AdminController_getUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar usuários com paginação e filtro",
        "tags": [
          "Admin - Users"
        ]
      },
      "post": {
        "operationId": "AdminController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar um novo usuário (USER ou ADMIN)",
        "tags": [
          "Admin - Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "2GO / AppRoteiros API",
    "description": "API de Roteiros de Viagem com IA e Autenticação OTP",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserAdminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "fullName": {
            "type": "string",
            "example": "João Silva"
          },
          "password": {
            "type": "string",
            "example": "senhaForte123",
            "minLength": 6
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ],
            "default": "USER"
          }
        },
        "required": [
          "email",
          "fullName",
          "password",
          "role"
        ]
      },
...
    }
  }
}